Refactor embedding service and repository: update API key handling, add vector search functionality, and enhance AI chat use case with language detection and translation features.

+++ repository/ai_chat_repository.go
@@ -37,7 +37,7 @@ func (r *AIUserChatRepositoryMongo) GetByUser(ctx context.Context, userID string
 		return nil, err
 	}
 	filter := bson.M{"user_id": userObjID}
-	opts := options.Find().SetSort(bson.D{{"timestamp", -1}}).SetLimit(int64(limit))
+	opts := options.Find().SetSort(bson.D{{"timestamp", int32(-1)}}).SetLimit(int64(limit))
 
 	cursor, err := r.collection.Find(ctx, filter, opts)
 	if err != nil {
